# app/docs/swagger.yaml
basePath: /v1
definitions:
...
  exception.InternalServerException:
    properties:
      code:
        example: 500
        type: integer
      message:
        example: internal server error
        type: string
    type: object
  exception.UnauthorizedException:
    properties:
      code:
        example: 401
        type: integer
      message:
        example: access not authorized
        type: string
    type: object
...
  handler.UpdateUserRequest:
    properties:
      active:
        example: true
        type: boolean
      email:
        example: johndoe@gmail.com
        type: string
      first_name:
        example: John
        type: string
      id:
        type: integer
      last_name:
        example: Doe
        type: string
      password:
        example: 12$a@3$@00!
        type: string
    type: object
host: localhost:8080
info:
  contact:
    name: Wesley Ferreira
    url: https://www.linkedin.com/in/wesleyfebarretos/
  description: Currency conversion API.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Challenge Bravo
  version: "1.0"
paths:
...
  /user/{id}:
    put:
      consumes:
      - application/json
      description: update user informing the id
      parameters:
      - description: new user data
        in: body
        name: newUser
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: boolean
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/exception.UnauthorizedException'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/exception.InternalServerException'
      security:
      - Bearer: []
      summary: Update User
      tags:
      - Users
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"